minor refector for project module in file-daemon
